change type to u32

+++ main.rs
@@ -1,6 +1,6 @@
 use std::env;
 
-fn fib(n: u64) -> u64 {
+fn fib(n: u32) -> u32 {
     return match n {
         0 => 0,
         1 => 1,
@@ -11,7 +11,7 @@ fn fib(n: u64) -> u64 {
 fn main() {
     let args: Vec<String> = env::args().collect();
 
-    let number: u64 = match args.len() {
+    let number: u32 = match args.len() {
         2 => match args[1].parse() {
             Ok(n) => n,
             Err(_) => {
